[move/#47] RegisterDropDownItem을 RegisterDropDown으로 이동
